% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_representation.R
\name{get_representation}
\alias{get_representation}
\title{GET the Representation of a Structure Identifier from CACTUS}
\usage{
get_representation(structure_identifier, representation, xml = FALSE)
}
\arguments{
\item{structure_identifier}{(Character.) The input; see Details.}

\item{representation}{(Character.) A character string indicating the desired
output format (NOT case sensitive); see Details.}

\item{xml}{(Logical.) Should the result be returned as XML string? Defaults
to \code{FALSE}.}
}
\value{
Returns the representation as (character) \code{vector}.
}
\description{
This function is used to GET the Representation of a structure
  identifier from the CACTUS Chemical Identifier Resolver API services.
}
\details{
If the input (\code{structure_identifier}) is an InChIKey, it needs
  to be prefixed with \code{"InChIKey="}, i.e.,
  \code{"InChIKey=RYYVLZVUVIJVGH-UHFFFAOYSA-N"} for caffeine.

  Allowed entries for \code{representation} are: \code{"StdInChI"},
  \code{"InChI"}, \code{"StdInChIKey"}, \code{"InChIKey"}, \code{"SMILES"},
  \code{"Names"} or \code{"IUPAC_Name"}.
}
\examples{
\dontrun{
## GET the standard InChI of caffeine
structure_identifier <- "caffeine"
representation <- "StdInChI"
get_representation(
  structure_identifier = structure_identifier,
  representation = representation
)
}
}
\seealso{
\url{https://cactus.nci.nih.gov/chemical/structure}
}
